+++ src/main/java/de/christophlorenz/jweatherflux/JWeatherFlux.java
@@ -1,0 +1,13 @@
+package de.christophlorenz.jweatherflux;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class JWeatherFlux {
+
+	public static void main(String[] args) {
+		SpringApplication.run(JWeatherFlux.class, args);
+	}
+
+}