+++ src/main/java/de/christophlorenz/jweatherflux/config/ForwarderProperties.java
@@ -1,7 +1,5 @@
-package de.christophlorenz.weatherflux.config;
+package de.christophlorenz.jweatherflux.config;
 
-import java.util.HashMap;
-import java.util.Map;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;