+++ src/main/java/de/christophlorenz/jweatherflux/controller/DataController.java
@@ -1,8 +1,8 @@
-package de.christophlorenz.weatherflux.controller;
+package de.christophlorenz.jweatherflux.controller;
 
-import de.christophlorenz.weatherflux.forwarder.Forwarder;
-import de.christophlorenz.weatherflux.service.DataReportService;
-import de.christophlorenz.weatherflux.service.PersistException;
+import de.christophlorenz.jweatherflux.forwarder.Forwarder;
+import de.christophlorenz.jweatherflux.service.DataReportService;
+import de.christophlorenz.jweatherflux.service.PersistException;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
@@ -11,8 +11,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.Mapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;