+++ src/main/java/de/christophlorenz/jweatherflux/model/WeatherData.java
@@ -1,4 +1,4 @@
-package de.christophlorenz.weatherflux.model;
+package de.christophlorenz.jweatherflux.model;
 
 import com.influxdb.client.domain.WritePrecision;
 import com.influxdb.client.write.Point;
@@ -6,10 +6,8 @@ import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 import java.util.Map;
-import java.util.TimeZone;
 import java.util.stream.Collectors;
 
 public class WeatherData {