+++ src/main/java/de/christophlorenz/jweatherflux/repository/InfluxRepositoryException.java
@@ -1,4 +1,4 @@
-package de.christophlorenz.weatherflux.repository;
+package de.christophlorenz.jweatherflux.repository;
 
 import com.influxdb.exceptions.InfluxException;