+++ src/test/java/de/christophlorenz/jweatherflux/JWeatherFluxTests.java
@@ -1,10 +1,10 @@
-package de.christophlorenz.weatherflux;
+package de.christophlorenz.jweatherflux;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class WeatherfluxApplicationTests {
+class JWeatherFluxTests {
 
 	@Test
 	void contextLoads() {